+++ packages/mui-components/src/components/atoms/Loader/index.ts
@@ -1,2 +1,3 @@
 export * from './Loader';
+export * from './Spinner';
 export * from './types';